feat: add encryption of CoseRecipients
