+++ src/token/cose/mod.rs
@@ -5,5 +5,6 @@ pub mod key;
 pub mod sign;
 
 pub mod mac;
+mod recipient;
 #[cfg(test)]
 mod test_helper;
